requirements.txt
addict==2.4.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
Markdown==3.4.1
markdown-it-py==2.2.0
matplotlib==3.7.0
mdurl==0.1.2
mmcls==0.25.0
mmcv-full==1.7.1
-e git+https://github.com/open-mmlab/mmsegmentation.git@c26915450564e9863980f40c2105c1b2a77a5f07#egg=mmsegmentation
model-index==0.1.11
mypy==1.0.1
mypy-extensions==1.0.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openmim==0.3.6
ordered-set==4.1.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
prettytable==3.6.0
Pygments==2.14.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rich==13.3.1
scipy==1.10.1
six==1.16.0
tabulate==0.9.0
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
types-Pillow==9.4.0.13
types-PyYAML==6.0.12.7
types-tqdm==4.64.7.13
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
yapf==0.32.0
